move core logic of convert_int_var_direct_encoding to direct::
